+++ src/common-pino-logger/common-pino.logger.ts
@@ -5,7 +5,7 @@ import { NodeEnvs } from '../enums/node-envs.enum';
 @Injectable()
 export class CommonPinoLogger extends PinoLogger {
   constructor(appName: string) {
-    console.log('creating logger...');
+    process.stdout.write('creating logger...');
 
     const targets = [];
     targets.push({
@@ -46,9 +46,11 @@ export class CommonPinoLogger extends PinoLogger {
         options: {
           colorize: true,
           colorizeObjects: true,
+          singleLine: true,
         },
       });
     }
+    console.log('done');
     super({
       pinoHttp: {
         name: appName,